extract attributes for export
